dh-autoreconf.pod: Document conflict with autotools-dev (Closes: #587102)
The following is the author's email.

Please do document the issue with autotools-dev. I'm starting to use your
package for packages where I already use autotools-dev addon, and if I didn't
happen to look at dh-autoreconf bug list, this bug would cause me to
unnecesarily waste time. OTOH If the manpage contained this information, it
would be a flawless experience, as I did read the manpage :-)

It seems more people will be in the same situation, so it's worthwile to
document this.