dh_autoreconf_clean: Correctly handle spaces in file names (Closes: #692317)
We split the checksums and filenames by simply using split, causing
us to ignore any part of the filename after a space. Limit the split
to split the line into two fields only.

This bug does not affect most projects, even if they have spaces
in their file names. It only affects projects where two files
in the same directory have a common prefix followed by a space,
and one of the files is modified (or the order they are visited
in changed).